Refactor: Refactor verify item components
