+++ react/admin/ReturnDetails/components/VerifyItems/AlignItemRight.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import type { FC } from 'react'
+
+export const AlignItemRight: FC = ({ children }) => (
+  <span className="flex justify-end w-100">{children}</span>
+)
